Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Amit,
> 
> I took a quick look at the drawings.   I did not see a drawing (or doc) that shows how Device Management and Voltha Core will interact.  For instance, for an image download, would NEM need to change the status of the device via Voltha first, then invoke the image download via Device Management, and when its completed then NEM change the device state?    In other words, are there any activities between Voltha and Device Management that NEM should coordinate?

Hi Khen,

There will be real interaction between Device Manager and vOLTHA. That is what I have tried to mention in the last para of the DeviceImageManagement.md file. Pasting it below

"In the case when an OLT device gets upgraded; for other external systems like vOLTHA, it would just be a reboot of the OLT and should be handled exactly the same way. ***If*** OLT/adapter/agent implementations store **presistent** config on the OLT then it needs to be ensured that this is available with the upgraded image so that vOLTHA does not see any difference compared to a normal reboot."

Let us know if you see any problems with this.

Patch-set: 1